+++ CHANGELOG.md
@@ -4,10 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.1] - 7.08.2021
+## [0.2.0] - 10.08.2021
+### Added
+- fromString() method to convert IEC types automatically from PLC data type declarations
+  - All non-complex data types supported (INT, REAL, ...)
+  - STRING and WSTRING support
+  - STRUCT support
+  - ARRAY support (multi-dimensional)
+
+### Updated
+- Updated multi-dimensional array handling to work properly
+- Updated README
+
+## [0.1.1] - 07.08.2021
 ### Added
 - README updates etc.
 
-## [0.1.0] - 7.08.2021
+## [0.1.0] - 07.08.2021
 ### Added
 - First release